exec: Update Operator interface to have Init return error
Because there is not an easy way without a large re-factor
to allow a pass over all operators in a flow, it seems that
the best way for memory monitoring of streaming operators
is to account for their memory upon initialization. This
means that operators need to be able to return errors
when their initialization would result in allocating too
much memory. Therefore, this PR performs the re-factor
of adding error returns to all the objects that
implement the Operator interface.

Release note: None