logictest: randomly apply primary key changes in logictests
This PR adds the ability for logic tests to randomly apply a primary
key change operation on create table statements that are within the
logic test. This PR does the following:

* On encountering a create table statement, the test runner computes
what columns are part of the primary key
* Generates a new index that indexes the same columns as the existing
primary key and stores all other columns in the table
* Generates an ALTER TABLE ALTER PRIMARY KEY command that switches the
primary key over to the new index.

TODOs:
* Figure out a way to disable this operation on a create table statement
* Figure out which tests that this operation should skip

Release note: None